% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DB-DB-API-files.R
\name{add_files}
\alias{add_files}
\title{Add a file to the database}
\usage{
add_files(path = character())
}
\arguments{
\item{path}{path to the file}
}
\value{
IDs of the added files in the database
}
\description{
This function is used to add files to the file paths database
}
