% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GUI-MODULE-add_update_delete_button.R
\name{add_update_delete_buttonUI}
\alias{add_update_delete_buttonUI}
\alias{add_update_delete_button}
\title{Module: Add / Update / Delete button control}
\usage{
add_update_delete_buttonUI(id)

add_update_delete_button(input, output, session, type, callback)
}
\arguments{
\item{id}{Namespace id parameter (must be unique)}

\item{input, output, session}{Shiny server parameters}

\item{type}{\code{\link[shiny]{reactive}} value to determine the state. 0 == add button, 1 == update and delete buttons, >1 == disabled update and enabled delete buttons}

\item{callback}{Function to call when a button is pressed}
}
\description{
Create an input button whose type depends upon a reactive value
}
\examples{
ui <- add_update_delete_buttonUI('myID')
server <- function(input, output, session) {
    callModule(add_update_delete_button, "myID",type = reactive(expr), callback = my_function)
}
}
