% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DO-arithmetic.R
\name{+.jms.data.object}
\alias{+.jms.data.object}
\alias{-.jms.data.object}
\alias{*.jms.data.object}
\alias{/.jms.data.object}
\title{Data object arithmetic}
\usage{
\method{+}{jms.data.object}(e1, e2)

\method{-}{jms.data.object}(e1, e2)

\method{*}{jms.data.object}(e1, e2)

\method{/}{jms.data.object}(e1, e2)
}
\arguments{
\item{e1}{The data object}

\item{e2}{The number(s) to add, subtract, multiply, or divide}
}
\value{
The data object
}
\description{
Perform arithmetic operations on the y columns for a data object.
\code{e2} can be:
\describe{
 \item{A single number}{The resulting data object has this number added / subtracted / ... to every y value}
 \item{A vector of length \code{nrow(e1)}}{e2 is added / subtracted / ... to each y column of e1}
 \item{A vector of length \code{length(ycol(e1))}}{Each y column has its corresponding value in e2 added / subtracted / ... to it}
 \item{A matrix of dimensions \code{nrow(e1) by length(ycol(e1))}}{Standard matrix addition / subtraction / ...}
}
}
