% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DO-definition.R
\name{as.jms.data.object}
\alias{as.jms.data.object}
\title{Convert an object into a JMS data object}
\usage{
as.jms.data.object(x)
}
\arguments{
\item{x}{The object to be converted}
}
\value{
The converted object
}
\description{
Convert an object into a JMS data object
}
