% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DB-definition.R
\name{as.jms.database}
\alias{as.jms.database}
\title{Convert an object into a jms.database}
\usage{
as.jms.database(x)
}
\arguments{
\item{x}{The object to be converted}
}
\value{
The converted object
}
\description{
Convert an object into a jms.database
}
