% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DB-TABLE-definition.R
\name{as.jms.database.table}
\alias{as.jms.database.table}
\title{Convert an object into a jms.database.table}
\usage{
as.jms.database.table(x, validator, version)
}
\arguments{
\item{x}{The object to be converted}
}
\value{
The converted object
}
\description{
Convert an object into a jms.database.table
}
