% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DB-DB-API-REACTIVE.R
\name{as.jms.reactive.database}
\alias{as.jms.reactive.database}
\title{Convert an object into a jms.database}
\usage{
as.jms.reactive.database(x, reactiveSession = NULL,
  reactiveUpdateFreq = 4000)
}
\arguments{
\item{x}{The object to be converted}
}
\value{
The converted object
}
\description{
Convert an object into a jms.database
}
