% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS-Type-Validation.R
\name{assert_colour}
\alias{assert_colour}
\title{Check whether or not a value is or can be coerced to a colour}
\usage{
assert_colour(colour, error)
}
\arguments{
\item{error}{The error message if the value is not a colour}

\item{value}{The value to test}
}
\value{
The value if it's a colour, error otherwise
}
\description{
Check whether or not a value is or can be coerced to a colour
}
