% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS-Type-Validation.R
\name{assert_date}
\alias{assert_date}
\title{Check whether or not a value is or can be coerced to a data}
\usage{
assert_date(value, error_msg)
}
\arguments{
\item{value}{The value to test}

\item{error_msg}{The error message if the value is not a ate}
}
\value{
The value if it's a date, error otherwise
}
\description{
Check whether or not a value is or can be coerced to a data
}
