% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS-Type-Validation.R
\name{assert_directory}
\alias{assert_directory}
\title{Check whether or not a directory exists}
\usage{
assert_directory(directory, error)
}
\arguments{
\item{directory}{The directory to test}

\item{error}{The error message if the directory does not exist}
}
\value{
The directory, error if it doesn't exist
}
\description{
Check whether or not a directory exists
}
