% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS-Type-Validation.R
\name{assert_function}
\alias{assert_function}
\title{Check whether or not a value is or can be coerced to a function}
\usage{
assert_function(func, error)
}
\arguments{
\item{error}{The error message if the value is not a function}

\item{value}{The value to test}
}
\value{
The value if it's a function, error otherwise
}
\description{
Check whether or not a value is or can be coerced to a function
}
