% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS-Type-Validation.R
\name{assert_list}
\alias{assert_list}
\title{Check whether or not a value is or can be coerced to a list}
\usage{
assert_list(value, error_msg)
}
\arguments{
\item{value}{The value to test}

\item{error_msg}{The error message if the value is not a list}
}
\value{
The value if it's a list, error otherwise
}
\description{
Check whether or not a value is or can be coerced to a list
}
