% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS-Type-Validation.R
\name{assert_logical}
\alias{assert_logical}
\title{Check whether or not a value is or can be coerced to a logial}
\usage{
assert_logical(value, error_msg = "Value must be logical")
}
\arguments{
\item{value}{The value to test}

\item{error_msg}{The error message if the value is not a logical}
}
\value{
The value if it's a logical, error otherwise
}
\description{
Check whether or not a value is or can be coerced to a logial
}
