% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS-packages.R
\name{assert_packages}
\alias{assert_packages}
\title{Raises an error if packages are not installed.}
\usage{
assert_packages(..., purpose = "This function")
}
\arguments{
\item{...}{Packages to test}

\item{purpose}{Used in error message}
}
\description{
Raises an error if packages are not installed.
}
