% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS-Type-Validation.R
\name{assert_positive}
\alias{assert_positive}
\title{Check whether or not a value is numeric or can be coerced to numeric
and is greater than or equal to 0}
\usage{
assert_positive(value, error_msg)
}
\arguments{
\item{value}{The value to test}

\item{error_msg}{The error message if the value is not positive}
}
\value{
The value if positive, error otherwise
}
\description{
Check whether or not a value is numeric or can be coerced to numeric
and is greater than or equal to 0
}
