% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GUI-Utils.R
\name{normal_button}
\alias{normal_button}
\alias{delete_button}
\alias{help_button}
\alias{info_button}
\title{Create a normal push button}
\usage{
normal_button(id, label, ...)

delete_button(id, ...)

help_button(id, ...)

info_button(id, ...)
}
\arguments{
\item{id}{The ID for the button}

\item{label}{The label for the button}

\item{...}{Further tags to set on the button}
}
\description{
Create a normal push button

Create a push button with the delete icon

Create a push button with the info icon

Create a push button with the help icon
}
