% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GUI-Utils.R
\name{checkbox}
\alias{checkbox}
\title{Create a checkbox input with support for extra tags}
\usage{
checkbox(id, ..., checked = TRUE)
}
\arguments{
\item{id}{The ID for the input}

\item{...}{Further tags to set on the input}
}
\description{
Create a checkbox input with support for extra tags
}
