% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DB-DB-API-files.R
\name{clear_files_database}
\alias{clear_files_database}
\title{Create an empty files database to replace the existing one}
\usage{
clear_files_database()
}
\description{
This function is used to initialise a database to hold details of file paths
}
