% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS-clipboard.R
\name{clipboard_paste}
\alias{clipboard_paste}
\alias{clipboard_copy}
\title{Get the contents of the clipboard}
\usage{
clipboard_paste()

clipboard_copy(string)
}
\description{
Get the contents of the clipboard

Copy data to the clipboard
}
