% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/METHODS-combine.R
\name{combine}
\alias{combine}
\title{Combine a list of data objects}
\usage{
combine(objects, interpolate)
}
\arguments{
\item{A}{\code{list} of jms.data.objects}
}
\value{
A jms.data.object containing the data
}
\description{
This function combines jms.data.objects
}
\examples{
combine(objects)
}
