% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GUI-MODULE-database_table.R
\name{database_table_UI}
\alias{database_table_UI}
\alias{database_table}
\title{Module: Display a database table}
\usage{
database_table_UI(id)

database_table(input, output, session, tableName, display_filter,
  buttons = NULL, database = project_database)
}
\arguments{
\item{id}{Namespace id parameter (must be unique)}

\item{input, output, session}{Shiny server parameters}

\item{tableName}{The name of the table in the database}

\item{display_filter}{Function to call to filter the data for display}

\item{buttons}{See \code{\link{addInputsToData}}}

\item{database}{The database, or a (possibly reactive) function that returns it}
}
\value{
\code{list} with a reactive element \code{count} giving the number of rows selected
a reactive \code{ids} to get the ids selected, a reactive \code{row} to get the row selected, and a vector \code{buttonIDs} of reactives to get the button ids pressed
}
\description{
Module: Display a database table
}
\examples{
ui <- database_table_UI('myID')
server <- function(input, output, session) {
    callModule(database_table, "myID",data_function = reactive(expr), display_filter = my_function)
}
}
