% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GUI-MODULE-database_table.R
\name{dataframe_table_UI}
\alias{dataframe_table_UI}
\alias{dataframe_table}
\title{Module: Display a dataframe table}
\usage{
dataframe_table_UI(id)

dataframe_table(input, output, session, dataframe, display_filter,
  buttons = NULL)
}
\arguments{
\item{id}{Namespace id parameter (must be unique)}

\item{input, output, session}{Shiny server parameters}

\item{dataframe}{A \code{\link[shiny]{reactive}} data.frame}

\item{display_filter}{Function to call to filter the data for display}

\item{buttons}{See \code{\link{addInputsToData}}}
}
\value{
\code{list} with a reactive element \code{count} giving the number of rows selected
a reactive \code{ids} to get the ids selected, a reactive \code{row} to get the row selected, and a vector \code{buttonIDs} of reactives to get the button ids pressed
}
\description{
Module: Display a dataframe table
}
\examples{
ui <- dataframe_table_UI('myID')
server <- function(input, output, session) {
    callModule(dataframe_table, "myID",data_function = reactive(expr), display_filter = my_function)
}
}
