% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DO-diff.R
\name{diff.jms.data.object}
\alias{diff.jms.data.object}
\title{Calculate Difference}
\usage{
\method{diff}{jms.data.object}(x, x2)
}
\arguments{
\item{x}{The 1st data object}

\item{x2}{The 2nd data object}
}
\value{
A data object representing the difference
}
\description{
This function calculates the difference between two data objects
}
\examples{
diff(x,x2)
}
