% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GUI-MODULE-dirChooser.R
\name{dirChooser}
\alias{dirChooser}
\alias{dirChooserUI}
\title{Module: Dir Chooser}
\usage{
dirChooser(input, output, session, state, updateFreq = 2000)

dirChooserUI(id, label, title, multiple, ..., buttonType = "default",
  class = NULL)
}
\arguments{
\item{input, output, session}{Shiny server parameters}

\item{state}{\code{\link[shiny]{reactive}} value to determine the state}
}
\value{
The value of the input
}
\description{
Calls either \code{\link{JS_dirChooser}} or \code{\link[rstudioapi]{selectDirectory}}
depending on whether it is used in an Rstudio instance
}
\examples{
ui <- dirChooserUI("myID", label="Select directory", title="Please select a directory")
server <- function(input, output, session) {
  callModule(dirChooser, "myID")
}
}
