% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GUI-Utils.R
\name{jsCodeHandler}
\alias{jsCodeHandler}
\alias{disableInput}
\alias{enableInput}
\alias{disableSelect}
\alias{enableSelect}
\title{Disable or enable an input}
\usage{
jsCodeHandler()

disableInput(id, session)

enableInput(id, session)

disableSelect(id, session)

enableSelect(id, session)
}
\arguments{
\item{id}{The ID for the input}

\item{session}{The session object}
}
\description{
Disable or enable an input

Disable or enable a select input
}
