% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DB-DB-API-files.R
\name{.blank_files_db}
\alias{.blank_files_db}
\title{Create an empty dataframe to function as the files database}
\usage{
.blank_files_db()
}
\description{
Create an empty dataframe to function as the files database
}
\keyword{internal}
