% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS-dygraphs-functions.R
\name{dyAxis.jms}
\alias{dyAxis.jms}
\title{Default axis for a \code{\link[dygraphs]{dygraph}}}
\usage{
dyAxis.jms(graph, name, label, valueRange, ticks = TRUE, digits = 2)
}
\arguments{
\item{name}{Axis name ('x', 'y', or 'y2')}

\item{label}{Label to display for axis (defaults to none).}

\item{valueRange}{Explicitly set the vertical range of the graph to 
\code{c(low, high)}. This may be set on a per-axis basis to define each 
y-axis separately. If either limit is unspecified, it will be calculated 
automatically (e.g. \code{c(NULL, 30)} to automatically calculate just the 
lower bound).}

\item{ticks}{Show ticks?}

\item{digits}{Number of decimal places in legend (only relevant if ticks==TRUE)}
}
\description{
Default axis for a \code{\link[dygraphs]{dygraph}}
}
