% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS-dygraphs-functions.R
\name{dyxlim}
\alias{dyxlim}
\title{Set the x limits for a \code{\link[dygraphs]{dygraph}}}
\usage{
dyxlim(dygraph, xlim)
}
\description{
Set the x limits for a \code{\link[dygraphs]{dygraph}}
}
