% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GUI-APP-API.R
\name{enable_gui_database_module}
\alias{enable_gui_database_module}
\alias{enable_gui_plot_module}
\alias{disable_gui_database_module}
\alias{disable_gui_plot_module}
\title{Enable or disable a GUI module}
\usage{
enable_gui_database_module(name)

enable_gui_plot_module(name)

disable_gui_database_module(name)

disable_gui_plot_module(name)
}
\arguments{
\item{name}{Title of the module}
}
\description{
Enable or disable a GUI module
}
