% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GUI-Utils.R
\name{NotificationStyle}
\alias{NotificationStyle}
\alias{show_error}
\alias{hide_error}
\title{Set the default notification style for errors and warnings to be centred on the screen}
\usage{
NotificationStyle()

show_error(content)

hide_error()
}
\arguments{
\item{content}{The content of the error message. Can be text / \code{\link[htmltools]{HTML}}}
}
\description{
This function is called to display an error to the user.
}
