% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PROCESSING-background.R
\name{expand_baseline_parameters}
\alias{expand_baseline_parameters}
\title{Expand a table of baseline parameters for use with \code{\link{make_backgrounds}}}
\usage{
expand_baseline_parameters(scans_and_points, nscans)
}
\arguments{
\item{scans_and_points}{data.frame with the 1st column containing lists of x points, the second column containing lists of scan numbers}
}
\value{
A list that can be used as the baseline_parameters argument to \code{\link{make_backgrounds}}
}
\description{
Expand a table of baseline parameters for use with \code{\link{make_backgrounds}}
}
