% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS-expression-convert.R
\name{expressionToHTML}
\alias{expressionToHTML}
\title{Convert an r \code{expression} to \code{HTML}}
\usage{
expressionToHTML(x)
}
\description{
Convert an r \code{expression} to \code{HTML}
}
\keyword{internal}
