% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS-expression-convert.R
\name{expressionToString}
\alias{expressionToString}
\title{Convert an r \code{expression} to a \code{character} string}
\usage{
expressionToString(x)
}
\description{
Convert an r \code{expression} to a \code{character} string
}
\keyword{internal}
