% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DB-DB-API-files.R
\name{file_for_ids}
\alias{file_for_ids}
\title{Get the file for an ID}
\usage{
file_for_ids(ids)
}
\arguments{
\item{ids}{id for one or more files}
}
\value{
The file paths
}
\description{
Get the file for an ID
}
