% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DB-DB-API-files.R
\name{files_db_validate}
\alias{files_db_validate}
\title{Validate entries to the files database}
\usage{
files_db_validate(path = character())
}
\arguments{
\item{path}{path to the file}
}
\description{
Validate entries to the files database
}
\keyword{internal}
