% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PACKAGE-onload.R
\name{get_configuration_directory}
\alias{get_configuration_directory}
\title{Create a system dependant configuration directory for use by a package}
\usage{
get_configuration_directory(package)
}
\description{
Create a system dependant configuration directory for use by a package
}
