% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GUI-APP-API.R
\name{gui_title}
\alias{gui_title}
\alias{set_gui_title}
\title{Get or set the main title for the GUI interface}
\usage{
gui_title()

set_gui_title(title)
}
\arguments{
\item{title}{The new title}
}
\description{
Get or set the main title for the GUI interface
}
