% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DB-DB-API-files.R
\name{id_for_files}
\alias{id_for_files}
\title{Get the ID for a file}
\usage{
id_for_files(paths)
}
\arguments{
\item{paths}{path to one or more files}
}
\value{
The ID(s) or NA if not found
}
\description{
Get the ID for a file
}
