% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GUI-Utils.R
\name{make_inputs}
\alias{make_inputs}
\alias{addInputsToData}
\alias{inputReactives}
\title{Make a vector of shiny inputs}
\usage{
make_inputs(input_id, ids, templateFunction, ..., session = NULL)

addInputsToData(data, inputs, session = NULL)

inputReactives(inputs, input)
}
\arguments{
\item{input_id}{The input that will be called when any input is clicked}

\item{ids}{Which id will be set for each button}

\item{templateFunction}{Function to make a button}

\item{inputs}{see details}
}
\value{
The vector of buttons

\code{list(data=<data with extra columns>, htmlCols=<column #s with HTML>)}
}
\description{
Make a vector of shiny inputs
}
\details{
One or more columns of inputs may be added to the table via the \code{inputs} parameter. \cr
It takes a \code{list} consisting of the following elements: \cr
\code{list( \cr
    name=character(),   <- column name for the input \cr
    id=character(),     <- input called when any input is clicked \cr
    template=function(),<- function that returns a input \cr
    column=numeric(),   <- column in which to put the input \cr
    [ids=data[,id]],    <- ids associated with the inputs (length = nrow(data)) \cr
    [... -- additional tags for the input])} \cr
 Buttons are generated using \code{make_inputs}.
}
\examples{
inputs=list(
list(
  name='test',
  id='checkbox',
  ids=rownames(data),
  template=checkbox,
  column=1
),
list(
  name='hr',
  id='',
  ids=rgb(t(col2rgb(plotrix::color.gradient(c(1,0),0,c(0,1),nslices=nrow(mydata)))/255)),
  template=hr,
  column=3
)
)
}
\keyword{internal}
