% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GUI-MODULE-basic_interactive_plot.R
\name{basicInteractivePlotUI}
\alias{basicInteractivePlotUI}
\alias{basicInteractivePlot}
\title{Module: Display a zoomable plot}
\usage{
basicInteractivePlotUI(id)

basicInteractivePlot(input, output, session, plotFunction, ...)
}
\arguments{
\item{id}{Namespace id parameter (must be unique)}

\item{input, output, session}{Shiny server parameters}

\item{plotFunction}{Function to produce a plot}

\item{...}{Extra parameters are passed to \code{plotFunction}}

\item{header}{Header for the legend}
}
\description{
Module: Display a zoomable plot
}
\examples{
ui <- basicInteractivePlotUI("myID")
server <- function(input, output, session) {
  callModule(basicInteractivePlot, "myID", plotFunction=myFunction)
}
}
