% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DO-iplot.R, R/METHODS-iplot.R
\name{iplot.jms.data.object}
\alias{iplot.jms.data.object}
\alias{iplot}
\alias{iplot.default}
\alias{iplotOutput}
\alias{renderIPlot}
\title{Make an interactive plot for a data object}
\usage{
\method{iplot}{jms.data.object}(..., offset = 1/sqrt(length(ycol(data)) - 1),
  xlim = NULL, ylim = NULL, y2lim = NULL, axes = c(1, 2),
  xlab = xlab_(data), ylab = ylab_(data), y2lab = y2lab_(data),
  col = graphics::par("col"), lwd = 1, pch = NA, labels = NULL,
  group = NULL, unsortedRangeWorkaround = T)

iplot(...)

\method{iplot}{default}(...)

iplotOutput(id)

renderIPlot(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{...}{further \link{graphical parameters} as in \code{\link{par}}.
    The relevant ones are \code{xaxs}, \code{yaxs} and \code{lab}.
  }

\item{xlim}{numeric vectors of length 2, giving the x and y
    coordinates ranges.}

\item{ylim}{numeric vectors of length 2, giving the x and y
    coordinates ranges.}

\item{axes}{a logical value indicating whether both axes should be drawn on
    the plot.  Use \link{graphical parameter} \code{"xaxt"} or \code{"yaxt"}
    to suppress just one of the axes.}

\item{xlab}{a label for the x axis, defaults to a description of \code{x}.}

\item{ylab}{a label for the y axis, defaults to a description of \code{y}.}

\item{col}{color code or name, see \code{\link{colors}},
    \code{\link{palette}}.  Here \code{NULL} means colour 0.}

\item{lwd}{line width, also used for (non-filled) plot symbols, see
    \code{\link{lines}} and \code{\link{points}}.}

\item{pch}{character or integer code for kind of points, see
    \code{\link{points.default}}.}

\item{group}{Group to associate this plot with.
The x-axis zoom level of plots within a group is automatically synchronized.}
}
\description{
Make an interactive plot for a data object
}
