% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS-Type-Validation.R
\name{is.Date}
\alias{is.Date}
\title{Check whether or not a value is a date}
\usage{
is.Date(x)
}
\arguments{
\item{x}{The value to test}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
Check whether or not a value is a date
}
