% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DO-definition.R
\name{is.jms.data.object}
\alias{is.jms.data.object}
\title{Check if an object is a jms.data.object}
\usage{
is.jms.data.object(x)
}
\arguments{
\item{x}{The object to be tested}
}
\value{
TRUE / FALSE
}
\description{
Check if an object is a jms.data.object
}
