% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DB-TABLE-ID-definition.R
\name{is.jms.database.table.id}
\alias{is.jms.database.table.id}
\title{Check if an object is a jms.database.table.id}
\usage{
is.jms.database.table.id(x)
}
\arguments{
\item{x}{The object to be tested}
}
\value{
TRUE / FALSE
}
\description{
Check if an object is a jms.database.table.id
}
