% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DB-DB-API-REACTIVE.R
\name{is.jms.reactive.database}
\alias{is.jms.reactive.database}
\title{Check if an object is a jms.reactive.database}
\usage{
is.jms.reactive.database(x)
}
\arguments{
\item{x}{The object to be tested}
}
\value{
TRUE / FALSE
}
\description{
Check if an object is a jms.reactive.database
}
