% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/METHODS-is.stale.R
\name{is.stale}
\alias{is.stale}
\title{Checks whether a database has been updated on disk, but not in memory}
\usage{
is.stale(x)
}
\arguments{
\item{x}{The database to check}
}
\description{
Checks whether a database has been updated on disk, but not in memory
}
