% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DB-DB-add_table.R
\name{[[<-.jms.database}
\alias{[[<-.jms.database}
\alias{[<-.jms.database}
\alias{$<-.jms.database}
\alias{+.jms.database}
\alias{-.jms.database}
\title{Add a table to a database}
\usage{
\method{[[}{jms.database}(x, name) <- value

\method{[}{jms.database}(x, name) <- value

\method{$}{jms.database}(x, name) <- value

\method{+}{jms.database}(e1, e2)

\method{-}{jms.database}(e1, e2)
}
\arguments{
\item{x}{The database}

\item{name}{The name of the table}

\item{value}{The table}
}
\value{
Database object
}
\description{
This function is used to add a \code{\link{jms.database.table}} to a database
}
\examples{
database=jms.database(<path>)
database[['mytable']]<-jms.database.table(...)
}
