% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DB-DB-API.R
\name{[[<-.jms.database}
\alias{[[<-.jms.database}
\alias{[[.jms.database}
\alias{[<-.jms.database}
\alias{$<-.jms.database}
\alias{[.jms.database}
\alias{$.jms.database}
\alias{+.jms.database}
\alias{-.jms.database}
\title{Add a table to a database}
\usage{
\method{[[}{jms.database}(x, name) <- value

\method{[[}{jms.database}(x, name)

\method{[}{jms.database}(x, name) <- value

\method{$}{jms.database}(x, name) <- value

\method{[}{jms.database}(x, name, ...)

\method{$}{jms.database}(x, name, ...)

\method{+}{jms.database}(e1, e2)

\method{-}{jms.database}(e1, e2)
}
\arguments{
\item{x}{The database}

\item{name}{The name of the table}

\item{value}{The table}
}
\value{
Database or table object
}
\description{
These functions are used to add or get a \code{\link{jms.database.table}} to or from a database
}
\examples{
\dontrun{
database <- jms.database("<path>")
database[["mytable"]] <- jms.database.table(...)
table <- database[["mytable"]]
}
}
