% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/METHODS-lock.R
\name{lock}
\alias{lock}
\alias{unlock}
\title{Lock a database to prevent editing}
\usage{
lock(x)

unlock(x, force = FALSE)
}
\arguments{
\item{x}{The database to lock}
}
\description{
Lock a database to prevent editing
}
