% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DB-definition.R
\name{jms.database}
\alias{jms.database}
\title{Create a database  object}
\usage{
jms.database(path = NULL)
}
\arguments{
\item{path}{Path to the directory containing the database on disk}
}
\value{
Database object
}
\description{
This function is used to initialise a database
}
\details{
A database object holds \code{\link{jms.database.table}} objects
}
