% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DB-TABLE-ID-definition.R
\name{jms.database.table.id}
\alias{jms.database.table.id}
\title{Create a database table object including support for an ID column}
\usage{
jms.database.table.id(..., validator = NULL)
}
\arguments{
\item{...}{parameters are passed to \code{\link{data.frame}}}

\item{version}{Can be used to track changes to table schema}
}
\value{
A database table object
}
\description{
This function is used to initialise a database table
}
