% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DB-TABLE-definition.R
\name{jms.database.table}
\alias{jms.database.table}
\title{Create a database table object}
\usage{
jms.database.table(..., version = 1)
}
\arguments{
\item{...}{parameters are passed to \code{\link{data.frame}}}

\item{version}{Can be used to track changes to table schema}
}
\value{
A database table object
}
\description{
This function is used to initialise a database table
}
