% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS-logging.R
\name{jms.enable.logging}
\alias{jms.enable.logging}
\alias{jms.logging.threshold}
\alias{jms.disable.logging}
\alias{jms.logging.file}
\alias{jms.logging.function}
\alias{log.trace}
\alias{log.debug}
\alias{log.info}
\alias{log.warn}
\alias{log.error}
\alias{log.fatal}
\title{Enable debug logging}
\usage{
jms.enable.logging(threshold)

jms.logging.threshold(threshold, ns = NULL)

jms.disable.logging()

jms.logging.file(file)

jms.logging.function(app)

log.trace(msg, ..., ns=NULL)

log.debug(msg, ..., ns=NULL)

log.info(msg, ..., ns=NULL)

log.warn(msg, ..., ns=NULL)

log.error(msg, ..., ns=NULL)

log.fatal(msg, ..., ns=NULL)
}
\arguments{
\item{threshold}{(optional for jms.enable.logging) see \code{\link[futile.logger]{flog.threshold}}}

\item{ns}{Logging namespace (see details)}
}
\description{
Enable debug logging
}
\details{
Enables logging output. Logging operates under a hierarchal set of namespaces, each
namespace can have its own threshold. Currently the predefined namespaces are:
\describe{
 \item{[package-name]}{The default logger is the name of the package which is doing the logging.
                       Note . is replaced by - in the name if necessary. (inherits from jms-logging)}
 \item{jms-database}{All database logging falls under this namespace (inherits from jms-logging)}
 \item{jms-database-table}{All database table logging falls under this namespace (inherits from jms-database)}
 \item{jms-logging}{The root logger for all packages developed by me}
}
}
