% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DB-DB-API-REACTIVE.R
\name{jms.reactive.database}
\alias{jms.reactive.database}
\title{Create a database  object}
\usage{
jms.reactive.database(path = NULL, reactiveSession = NULL,
  reactiveUpdateFreq = 4000)
}
\arguments{
\item{path}{Path to the directory containing the database on disk}

\item{reactiveSession}{The \code{shiny} session object}

\item{reactiveUpdateFreq}{Approximate number of milliseconds to wait between polling for updates.}
}
\value{
Database object
}
\description{
This function is used to initialise a reactive database
}
\details{
A database object holds \code{\link{jms.database.table}} objects
}
