% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GUI-MODULE-layoutInput.R
\name{layoutInput}
\alias{layoutInput}
\title{Module: layout input}
\usage{
layoutInput(id, default_layout, plot_names, can_add_plots, min_plots)
}
\description{
Create an input control that allows a user to determine a plot layout. \cr
Returns a list of nPlots: <int>, matrix: <matrix>, input: <input for default_layout>\cr
The matrix can be used with the \code{\link{layout}} command.
}
