% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DO-plot.R
\name{lines.jms.data.object}
\alias{lines.jms.data.object}
\title{Draw lines for a data object}
\usage{
\method{lines}{jms.data.object}(x, col = graphics::par("col"), type = "l",
  y2 = TRUE, ...)
}
\arguments{
\item{x}{The object to plot}

\item{col}{color code or name, see \code{\link{colors}},
    \code{\link{palette}}.  Here \code{NULL} means colour 0.}

\item{type}{1 character code: see \code{\link{plot.default}}.
    \code{NULL} is accepted as a synonym for \code{"p"}.}

\item{y2}{Include y2 data in plot?}

\item{...}{further \link{graphical parameters} such as \code{xpd},
    \code{lend}, \code{ljoin} and \code{lmitre}.}
}
\description{
This function plots a data object
}
\examples{
plot.xy(data)
}
