% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS-list.files.sorted.R
\name{list.files.sorted}
\alias{list.files.sorted}
\title{list.files with sorting}
\usage{
list.files.sorted(...)
}
\arguments{
\item{...}{Parameters passed to \code{\link{list.files}}}
}
\description{
list.files with sorting
}
\section{Sorting}{

Natural sorting will take a file name of the sort <string><number><string>.<extension> and sort based on the number in a natural manner that does not depend on the number of digits in that number.
i.e 10 would come after 9 in a natural sort, but after 1 using the default sorting method
}

