% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DO-load.R
\name{load.file}
\alias{load.file}
\title{Read a data file}
\usage{
\method{load}{file}(path, func, ...)
}
\arguments{
\item{path}{The path to the file / directory}

\item{func}{The function to read the file}
}
\value{
A jms.data.object containing the data
}
\description{
This function reads one data file
}
\examples{
load.file('/path/to/file.ext', load_function)
}
\keyword{internal}
