% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DO-load.R
\name{load.jms}
\alias{load.jms}
\title{Read data files}
\usage{
load.jms(path, func, ext = NULL, pattern = NULL, sort = FALSE, ...)
}
\arguments{
\item{path}{The path to the file / directory}

\item{func}{The function to read the file}

\item{ext}{The file extension to look for}

\item{pattern}{The pattern to look for}

\item{sort}{If TRUE then use a "natural" numerical sort (see \code{\link[Plotting.Utils]{list.files.sorted}}), else use the default alphabetical sort}

\item{...}{Additional parameters are passed to func}
}
\value{
A jms.data.object containing the data
}
\description{
This function reads one or more data files in a directory
}
\examples{
load.jms('/path/to/directory', load_function, ext='ext')
load.jms('/path/to/file.ext', load_function)
load.jms(c('/path/to/file.ext','/path/to/file2.ext'))
}
