% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS-lockfile.R
\name{make_lockfile}
\alias{make_lockfile}
\alias{remove_lockfile}
\title{Create a directory to act as a lockfile, error if the file already exists}
\usage{
make_lockfile(path, timeout = 10)

remove_lockfile(path)
}
\arguments{
\item{path}{Path to the lockfile}

\item{timeout}{How many times to attempt to obtain the lock}
}
\description{
Create a directory to act as a lockfile, error if the file already exists
}
