% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PROCESSING-background.R
\name{make_background}
\alias{make_background}
\title{Create a background based on a vector of x values which lie on the baseline}
\usage{
make_background(xy, x_points, bkg_y_avg_points = 4, returnFunc = FALSE)
}
\arguments{
\item{xy}{The data to create the background for (data.frame(x=..., y=...))}

\item{x_points}{X values which lie on the baseline}

\item{bkg_y_avg_points}{How many points either side of each x point to use to account for noise}

\item{returnFunc}{Returns an splinefun rather than a vector of y values}
}
\value{
A vector containing the y values for the background
}
\description{
This function creates a spline background for the given data and x points
}
