% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PROCESSING-background.R
\name{make_backgrounds}
\alias{make_backgrounds}
\title{Create backgrounds based on a list of vectors of x values which lie on the baseline}
\usage{
make_backgrounds(data, baseline_parameters)
}
\arguments{
\item{data}{The data to create the background for (data.frame(x=..., y=..., y2=...))}

\item{baseline_parameters}{list of length equal to the number of y columns.
A background will be created for each element of the list using
\code{\link{make_background}}.}
}
\value{
A \code{jms.data.object} containing the backgrounds that can be subtracted from the data
}
\description{
This function creates a spline background for the given data and x points
}
