% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS-match.names.R
\name{match.names}
\alias{match.names}
\title{match the items in a list to a vector of names}
\usage{
match.names(names, list)
}
\arguments{
\item{names}{vector of names to be expected from list}

\item{list}{list of values to be matched with names}
}
\value{
list with items ordered and names as defined by another vector
}
\description{
match.names returns a list with items ordered and names as defined by another vector
}
