% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DO-range.R
\name{max.jms.data.object}
\alias{max.jms.data.object}
\title{Calculate Max}
\usage{
\method{max}{jms.data.object}(x, ...)
}
\arguments{
\item{x}{The data object}
}
\value{
The maximum y value
}
\description{
This function calculates the minimum y value of a data object
}
\examples{
max(data)
}
