% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DO-range.R
\name{min.jms.data.object}
\alias{min.jms.data.object}
\title{Calculate Min}
\usage{
\method{min}{jms.data.object}(x, ...)
}
\arguments{
\item{x}{The data object}
}
\value{
The minimum y value
}
\description{
This function calculates the minimum y value of a data object
}
\examples{
min(data)
}
