% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS-collapse_numeric_ranges.R
\name{numeric_to_string_ranges}
\alias{numeric_to_string_ranges}
\title{Collapse a numeric vector to a character vector of ranges}
\usage{
numeric_to_string_ranges(x, min_collapse = 2)
}
\arguments{
\item{x}{The numeric vector}

\item{min_collapse}{Minimum length of range before collapsing}
}
\description{
Collapse a numeric vector to a character vector of ranges
}
\examples{
s <- c(1, 2, 3, 4, 8, 9, 14, 15, 16, 19, 21, 24, 25, 26)

paste0(numeric_to_string_ranges(s), collapse=", ")
paste0(numeric_to_string_ranges(s, 3), collapse=", ")
paste0(numeric_to_string_ranges(s, 4), collapse=", ")
}
