% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS-paths.R
\name{cannonicalPath}
\alias{cannonicalPath}
\alias{clipboard_to_path}
\title{Get the cannonical path to a file}
\usage{
cannonicalPath(path)

clipboard_to_path()
}
\arguments{
\item{path}{Path to the file}
}
\description{
Get the cannonical path to a file

Get the cannonical path to a file using data on the clipboard
}
