% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DB-project.R
\name{project_database}
\alias{project_database}
\alias{setproject_database}
\title{Set or get the current project database}
\usage{
project_database()

setproject_database(db)
}
\description{
Set or get the current project database
}
