% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DO-range.R
\name{range.jms.data.object}
\alias{range.jms.data.object}
\title{Calculate Range}
\usage{
\method{range}{jms.data.object}(x, offset = 0, ...)
}
\arguments{
\item{x}{The data object}
}
\value{
The range
}
\description{
This function calculates the range of a data objects
}
\examples{
range(data)
}
