% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DO-definition.R
\name{read.table.jms}
\alias{read.table.jms}
\title{Read a table as a JMS data object}
\usage{
read.table.jms(...)
}
\arguments{
\item{...}{parameters are passed to \code{\link{read.table}}}
}
\value{
A JMS data object containing the data
}
\description{
Read a table as a JMS data object
}
