% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GUI-APP-API.R
\name{register_gui_database_module}
\alias{register_gui_database_module}
\title{Register a module to form a tab in the GUI}
\usage{
register_gui_database_module(moduleUI, moduleServer, title,
  databases = c(), namespace = NA)
}
\arguments{
\item{moduleUI}{character string representing the name of the ui function in namespace}

\item{moduleServer}{character string representing the name of the server function in namespace}

\item{title}{The (user visible) title of the module}

\item{databases}{The names of the databases (part of \code{\link{project_database}}) that the module requires access to}

\item{namespace}{The namespace of the ui and server functions}
}
\description{
Register a module to form a tab in the GUI
}
\details{
Modules have a UI and a server function \cr
The UI function takes a single parameter \code{id} as per the shiny module documentation \cr
The server function takes 3+ parameters, \code{input}, \code{output} and \code{session} plus reactives
corresponding to the databases requested. \cr
Server functions should return a list with the following elements:
\itemize{
 \item{"name"}{The name of the token referencing this module's data (see \code{\link{register_gui_database_module}})}
 \item{"plotbutton"}{A \code{\link[shiny]{reactive}} referencing the plot button on the page}
 \item{"ids"}{A \code{\link[shiny]{reactive}} list of ids to plot}
}
This function should usually be used within a package's \code{.onLoad} function.
}
