% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GUI-APP-API.R
\name{register_gui_plot_module}
\alias{register_gui_plot_module}
\title{Register a module to form a plot tab in the GUI}
\usage{
register_gui_plot_module(moduleUI, moduleServer, title, tokens = c(),
  namespace = NA)
}
\arguments{
\item{moduleUI}{character string representing the name of the ui function in namespace}

\item{moduleServer}{character string representing the name of the server function in namespace}

\item{title}{The (user visible) title of the module}

\item{namespace}{The namespace of the ui and server functions}

\item{token}{Names for tokens which represent the data plotted by this module}
}
\description{
Register a module to form a plot tab in the GUI
}
\details{
Modules have a UI and a server function \cr
The UI function takes a single parameter \code{id} as per the shiny module documentation \cr
The server function takes 5 parameters, \code{input}, \code{output} and \code{session} plus
an \code{ids} reactive representing the \code{ids} to plot and an errorFunction to be called
with a message in case of an error. \cr
This function should usually be used within a package's \code{.onLoad} function.
}
