% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DB-DB-API-files.R
\name{remove_files}
\alias{remove_files}
\title{Remove files from the database}
\usage{
remove_files(ids)
}
\arguments{
\item{ids}{IDs of the files}
}
\description{
This function is used to remove files from the files database
}
