% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GUI-MODULE-dirChooser.R
\name{rstudio_dirChooser}
\alias{rstudio_dirChooser}
\title{Module: RStudio Dir Chooser}
\usage{
rstudio_dirChooser(input, output, session, state)
}
\arguments{
\item{input, output, session}{Shiny server parameters}

\item{state}{\code{\link[shiny]{reactive}} value to determine the state}
}
\value{
The value of the input
}
\description{
Module: RStudio Dir Chooser
}
\examples{
ui <- rstudio_dirChooserUI('myID', label='Select Directory')
server <- function(input, output, session) {
    callModule(rstudio_dirChooser, "myID")
}
}
