% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GUI-MODULE-fileChooser.R
\name{rstudio_fileChooser}
\alias{rstudio_fileChooser}
\alias{rstudio_fileChooserUI}
\title{Module: RStudio File Chooser}
\usage{
rstudio_fileChooser(input, output, session, state, filetypes = NULL)

rstudio_fileChooserUI(id, label, ..., buttonType = "default", class = NULL)
}
\arguments{
\item{input, output, session}{Shiny server parameters}

\item{state}{\code{\link[shiny]{reactive}} value to determine the state}
}
\value{
The value of the input
}
\description{
Module: RStudio File Chooser
}
\examples{
ui <- rstudio_fileChooserUI('myID', label='File select')
server <- function(input, output, session) {
    rstudio_fileChooser(rstudio_fileChooser, "myID")
}
}
