% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DB-TABLE-ID-get.R
\name{[.jms.database.table.id}
\alias{[.jms.database.table.id}
\title{Get an element from the table}
\usage{
\method{[}{jms.database.table.id}(x, i = T, j = T)
}
\description{
Get an element from the table
}
