% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DB-get-table.R
\name{[[.jms.database}
\alias{[[.jms.database}
\title{Get a database table from a database}
\usage{
\method{[[}{jms.database}(x, name, ..., reactiveSession = NULL,
  reactiveUpdateFreq = 4000)
}
\description{
Get a database table from a database
}
