% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DB-TABLE-ID-update.R
\name{[<-.jms.database.table.id}
\alias{[<-.jms.database.table.id}
\title{Update a row in the table}
\usage{
\method{[}{jms.database.table.id}(x, i = NULL, j = NULL) <- value
}
\description{
Update a row in the table
}
