% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GUI-MODULE-toggleInput.R
\name{toggleInput}
\alias{toggleInput}
\title{Like a checkbox input, but displays as a toggle}
\usage{
toggleInput(inputId, label, value = FALSE, onlabel = "",
  offlabel = "", style = NULL)
}
\description{
Like a checkbox input, but displays as a toggle
}
