% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DO-undersample.R
\name{undersample}
\alias{undersample}
\title{Undersamples a large data object to speed up processing / plotting etc.}
\usage{
undersample(data, n)
}
\arguments{
\item{data}{The data object}

\item{n}{For every n points, only the first will be taken}
}
\value{
The undersampled data object
}
\description{
Undersamples a large data object to speed up processing / plotting etc.
}
