% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DB-DB-API-files.R
\name{update_files}
\alias{update_files}
\title{Update a file in the database}
\usage{
update_files(id = numeric(), path = character())
}
\arguments{
\item{id}{ID of the file to update}

\item{path}{path to the file}
}
\value{
ID
}
\description{
This function is used to update a file in the files database
}
