% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS-Type-Validation.R
\name{verify_positive}
\alias{verify_positive}
\title{Check whether or not a value is numeric or can be coerced to numeric
and is greater than or equal to 0}
\usage{
verify_positive(value, if_empty = TRUE)
}
\arguments{
\item{value}{The value to test}

\item{if_empty}{The value to return if length(value)==0}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
Check whether or not a value is numeric or can be coerced to numeric
and is greater than or equal to 0
}
