% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/METHODS-xylab.R
\name{xlab}
\alias{xlab}
\alias{xlab<-}
\alias{ylab}
\alias{ylab<-}
\alias{y2lab}
\alias{y2lab<-}
\title{Get or set the x or y column labels for a data object}
\usage{
xlab(x)

xlab(x) <- value

ylab(x)

ylab(x) <- value

y2lab(x)

y2lab(x) <- value
}
\arguments{
\item{x}{The data object}
}
\value{
The column labels
}
\description{
Get or set the x or y column labels for a data object
}
