% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/METHODS-xyscale.R
\name{xscale}
\alias{xscale}
\alias{xscale<-}
\alias{yscale}
\alias{yscale<-}
\alias{y2scale}
\alias{y2scale<-}
\title{Get or set the x or y column scale factors for a data object}
\usage{
xscale(x)

xscale(x) <- value

yscale(x)

yscale(x) <- value

y2scale(x)

y2scale(x) <- value
}
\arguments{
\item{x}{The data object}
}
\value{
The column scales
}
\description{
Get or set the x or y column scale factors for a data object
}
