% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DO-arithmetic.R
\name{+.jms.data.object}
\alias{+.jms.data.object}
\alias{-.jms.data.object}
\alias{*.jms.data.object}
\alias{/.jms.data.object}
\title{Add a number to a data objects y values}
\usage{
\method{+}{jms.data.object}(e1, e2)

\method{-}{jms.data.object}(e1, e2)

\method{*}{jms.data.object}(e1, e2)

\method{/}{jms.data.object}(e1, e2)
}
\arguments{
\item{e1}{The data object}

\item{e2}{The number to add}

\item{e1}{The data object}

\item{e2}{The number to subtract}

\item{e1}{The data object}

\item{e2}{The number by which to multiply}

\item{e1}{The data object}

\item{e2}{The number by which to divide}
}
\value{
The data object

The data object

The data object

The data object
}
\description{
Add a number to a data objects y values

Subtract a number to a data objects y values

Multiply a data objects y values by a number

Divide a data objects y values by a number
}
