% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GUI-Utils.R
\name{show_error}
\alias{show_error}
\alias{hide_error}
\title{Present an error dialog to the user}
\usage{
show_error(e, errorid, session)

hide_error(errorid, session)
}
\arguments{
\item{e}{The error text}

\item{errorid}{The ID of the error dialog box}

\item{session}{The session object}
}
\description{
This function is called to display an error to the user.
}
\keyword{internal}
