% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DO-definition.R
\name{jms.data.object}
\alias{jms.data.object}
\title{Make a new JMS data object}
\usage{
jms.data.object(...)
}
\arguments{
\item{...}{parameters are passed to \code{\link{data.frame}}}
}
\value{
A JMS data object
}
\description{
Make a new JMS data object
}
