% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DB-TABLE-definition.R
\name{jms.database.table}
\alias{jms.database.table}
\title{Create a database table object}
\usage{
jms.database.table(...)
}
\arguments{
\item{...}{parameters are passed to \code{\link{data.frame}}}
}
\value{
A database table object
}
\description{
This function is used to initialise a database table
}
\keyword{internal}
