% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS-logging.R
\name{jms.enable.logging}
\alias{jms.enable.logging}
\alias{jms.logging.setnamespace}
\alias{jms.logging.threshold}
\alias{jms.disable.logging}
\alias{jms.logging.file}
\alias{jms.logging.function}
\alias{log.trace}
\alias{log.debug}
\alias{log.info}
\alias{log.warn}
\alias{log.error}
\alias{log.fatal}
\title{Enable debug logging}
\usage{
jms.enable.logging(threshold)

jms.logging.setnamespace(name)

jms.logging.threshold(threshold, ns = NULL)

jms.disable.logging()

jms.logging.file(file)

jms.logging.function(app)

\method{log}{trace}(msg, ...)

\method{log}{debug}(msg, ...)

\method{log}{info}(msg, ...)

\method{log}{warn}(msg, ...)

\method{log}{error}(msg, ...)

\method{log}{fatal}(msg, ...)
}
\arguments{
\item{threshold}{(optional for jms.enable.logging) see \code{\link[futile.logger]{flog.threshold}}}
}
\description{
Enable debug logging
}
