% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DO-plot.R
\name{plot.jms.data.object}
\alias{plot.jms.data.object}
\title{Plot a data object}
\usage{
\method{plot}{jms.data.object}(x, offset = 1/sqrt(length(ycol(x)) - 1),
  xlim = NULL, ylim = NULL, y2lim = NULL, xlab = xlab_(x),
  ylab = ylab_(x), y2lab = y2lab_(x), axes = c(1, 2), ...)
}
\arguments{
\item{x}{The object to plot}

\item{offset}{For 2D data objects, the fraction of the yrange to offset the data}
}
\description{
This function plots a data object
}
\examples{
plot(data)
}
