% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DB-DB-API-project.R
\name{project_database}
\alias{project_database}
\alias{set_project_database}
\title{Get or set the current project database.}
\usage{
project_database()

set_project_database(db)
}
\arguments{
\item{db}{A \code{\link{jms.database}}}
}
\description{
Get or set the current project database. This setting is persistent across sessions.
}
