% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz-DB-project.R
\name{project_database}
\alias{project_database}
\alias{set_project_database}
\title{Set or get the current project database}
\usage{
project_database()

set_project_database(db)
}
\description{
Set or get the current project database
}
